今日するコンポーネントのリファクタ
