+++ src/components/TodaysToDo/TodaysToDoList.jsx
@@ -1,0 +1,17 @@
+import React from "react";
+import { Button } from ".";
+
+export const TodaysToDoList = ({ task, setIsOpenEdit }) => {
+  return (
+    <Button
+      type="button"
+      disabled={task.completed}
+      onClick={() => setIsOpenEdit(task.id)}
+      className={`text-left ${
+        task.completed ? "text-gray-400 line-through" : "text-black"
+      }`}
+    >
+      {task.task}
+    </Button>
+  );
+};
