+++ src/pages/index.jsx
@@ -2,12 +2,14 @@ import React from "react";
 import { TodaysToDo } from "src/components/TodaysToDo";
 import { Layout } from "src/layouts/Layout";
 
-export default function Home() {
-	return (
-		<Layout>
-			<div>
-				<TodaysToDo />
-			</div>
-		</Layout>
-	);
-}
+const Home = () => {
+  return (
+    <Layout>
+      <div>
+        <TodaysToDo />
+      </div>
+    </Layout>
+  );
+};
+
+export default Home;
